backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Data Analysis
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# LLM
google-generativeai==0.8.3

# Validation & Settings
pydantic==2.8.2
pydantic-settings==2.4.0

# Environment
python-dotenv==1.0.1